create file for all classes within acados_ocp
